+++ docs/source/conf.py
@@ -98,12 +98,12 @@ rst_epilog = """
 .. |year| replace:: {year}
 """.format(year=year)
 
+html_static_path = ["_static"]
 html_last_updated_fmt = '%b %d, %Y'
 html_title = 'PyProximal'
 html_short_title = 'PyProximal'
-html_logo = '_static/pyproximal.png'
-html_favicon = '_static/favicon.ico'
-html_static_path = ['_static']
+html_logo = 'pyproximal.png'
+html_favicon = 'favicon.ico'
 html_extra_path = []
 pygments_style = 'default'
 add_function_parentheses = False
@@ -112,7 +112,7 @@ html_show_sphinx = True
 html_show_copyright = True
 
 # Theme config
-html_theme = "sphinx_rtd_theme"
+html_theme = "pydata_sphinx_theme"
 html_theme_options = {
     'logo_only': True,
     'display_version': True,