+++ pyproximal/projection/AffineSet.py
@@ -1,5 +1,6 @@
-import numpy as np
-from scipy.sparse.linalg import lsqr
+from scipy.sparse.linalg import lsqr as sp_lsqr
+from pylops.optimization.basic import lsqr
+from pylops.utils.backend import get_array_module, get_module_name
 
 
 class AffineSetProj():
@@ -39,6 +40,9 @@ class AffineSetProj():
         self.niter = niter
 
     def __call__(self, x):
-        inv = lsqr(self.Op * self.Op.H, self.Op * x - self.b, iter_lim=self.niter)[0]
-        y = x - self.Op.H * inv
+        if get_module_name(get_array_module(x)) == 'numpy':
+            inv = sp_lsqr(self.Op * self.Op.H, self.Op * x - self.b, iter_lim=self.niter)[0]
+        else:
+            inv = lsqr(self.Op * self.Op.H, self.Op * x - self.b, niter=self.niter)[0]
+        y = x - self.Op.H * inv.ravel() # currently ravel is added to ensure that the output is always a vector
         return y