+++ pyproximal/proximal/L2.py
@@ -1,7 +1,10 @@
 import numpy as np
 from scipy.linalg import cho_factor, cho_solve
-from scipy.sparse.linalg import lsqr
+from scipy.sparse.linalg import lsqr as sp_lsqr
 from pylops import MatrixMult, Identity
+from pylops.optimization.basic import lsqr
+from pylops.utils.backend import get_array_module, get_module_name
+
 from pyproximal.ProxOperator import _check_tau
 from pyproximal import ProxOperator
 
@@ -45,6 +48,11 @@ class L2(ProxOperator):
         avoids to do so unless the :math:`\tau` or :math:`\sigma` paramets
         have changed. Choose ``densesolver=None`` when using PyLops versions
         earlier than v1.18.1 or v2.0.0
+    **kwargs_solver : :obj:`dict`, optional
+        Dictionary containing extra arguments for
+        :py:func:`scipy.sparse.linalg.lsqr` solver when using
+        numpy data (or :py:func:`pylops.optimization.solver.lsqr` and
+        when using cupy data)
 
     Notes
     -----
@@ -86,7 +94,8 @@ class L2(ProxOperator):
 
     """
     def __init__(self, Op=None, b=None, q=None, sigma=1., alpha=1.,
-                 qgrad=True, niter=10, x0=None, warm=True, densesolver=None):
+                 qgrad=True, niter=10, x0=None, warm=True,
+                 densesolver=None, kwargs_solver=None):
         super().__init__(Op, True)
         self.b = b
         self.q = q
@@ -98,6 +107,7 @@ class L2(ProxOperator):
         self.warm = warm
         self.densesolver = densesolver
         self.count = 0
+        self.kwargs_solver = {} if kwargs_solver is None else kwargs_solver
 
         # when using factorize, store the first tau*sigma=0 so that the
         # first time it will be recomputed (as tau cannot be 0)
@@ -164,8 +174,13 @@ class L2(ProxOperator):
                     x = cho_solve(self.cl, y)
             else:
                 Op1 = Identity(self.Op.shape[1], dtype=self.Op.dtype) + \
-                      tau * self.sigma * self.Op.H * self.Op
-                x = lsqr(Op1, y, iter_lim=niter, x0=self.x0)[0]
+                      float(tau * self.sigma) * (self.Op.H * self.Op)
+                if get_module_name(get_array_module(x)) == 'numpy':
+                    x = sp_lsqr(Op1, y, iter_lim=niter, x0=self.x0,
+                                **self.kwargs_solver)[0]
+                else:
+                    x = lsqr(Op1, y, niter=niter, x0=self.x0,
+                             **self.kwargs_solver)[0].ravel()
             if self.warm:
                 self.x0 = x
         elif self.b is not None: